custom snipmate folder
snipMate plays quite nicely with both its own folder in the bundle, and
this custom one under ~/.vim/snippets.  Just be sure not to name anything
the same as snippets in the filetype that you're working with (i.e. you
can't do a 'fun' or a 'for' as those snippets already exist in the bundle).