Made changes to vundle to make a gui version
(Maybe not necessary?)

The gui version contains color schemes only, the other version has "normal" plugins.

There were also some changes to plugins that were made (easytag went byebye, closetag
entered the picture).

The gui version does not try to set up vundle or make sure it is working.  It is
guaranteed to load later than the non-gui version because it is sourced by gvimrc
and the other version is sourced by vimrc[1].

[1] http://stackoverflow.com/questions/6977043/a-way-to-try-out-different-peoples-github-vim-setups